refactor: ErrorLayout 추가
- 에러 화면은 여기서 처리를 해줌
