+++ src/components/layout/ErrorLayout/index.ts
@@ -1,0 +1,1 @@
+export * from './ErrorLayout';
